Added owned cars to the query

+++ Data_Structures/Game.cs
@@ -11,7 +11,7 @@ namespace Thesis_backend.Data_Structures
         public User User { get; set; }
         public long CurrentXP { get; set; }
         public int NextLVLXP { get; set; }
-        public long Currency { get; set; }
+        public int Currency { get; set; }
         public List<OwnedCar>? OwnedCars { get; set; }
 
         public override object Serialize => new { ID, Lvl, CurrentXP, NextLVLXP, UserId, Currency, OwnedCars = OwnedCars?.Select(car => car.Serialize).ToList() };
